chore: reorganize types
