+++ gui/src/routes/+page.svelte
@@ -2,7 +2,7 @@
   import * as Sidebar from "$lib/components/ui/sidebar";
   import * as Dialog from "$lib/components/ui/dialog";
 
-  import type { UpdateInfo } from "../../../src/main/update";
+  import type { UpdateInfo } from "$sharedTypes/update";
   import UpdateIcon from "@lucide/svelte/icons/circle-fading-arrow-up";
   import RefreshCw from "@lucide/svelte/icons/refresh-cw";
