+++ src/main/menu/skin.ts
@@ -2,7 +2,7 @@ import photon from "@silvia-odwyer/photon-node";
 
 import packManager from "../pack";
 import SkinPack from "../packs/SkinPack";
-import type { MenuSkin } from "./types";
+import type { MenuSkin } from "$sharedTypes/menu";
 import { extractColor } from "../skin/color";
 import { argbToCss } from "../skin/dui";
