+++ src/main/menu/types.ts
@@ -1,34 +1,4 @@
-import { BtnImages } from "../../../types/dui";
-
-export type AppMenuItemBtn = {
-  id: string;
-  url: string;
-  images?: BtnImages | null;
-  enable: boolean;
-};
-
-export type AppMenuItem = {
-  text: string;
-  menu: boolean;
-  enable: boolean;
-  separator: boolean;
-  children: AppMenuItem[] | null;
-  hotkey?: string;
-  image_color: string;
-  image_path?: string;
-  check_image_path?: string;
-  menu_id: string | null;
-  style?: string;
-  btns?: AppMenuItemBtn[];
-};
-
-export type MenuSkin = {
-  background: string;
-  foreground: string;
-  foregroundDisabled: string;
-  separator: string;
-  itemHover: string;
-};
+import type { AppMenuItem } from "$sharedTypes/menu";
 
 export type MenuClickHandler = (menuId: string | null) => void;
