+++ src/preload.ts
@@ -6,6 +6,8 @@ import { contextBridge } from "electron";
 import "./preload/channel";
 import "./preload/desktopLyrics";
 
+import "./preload/calls/index";
+
 contextBridge.executeInMainWorld({
   func: () => {
     const originalContentDocumentDescriptor = Object.getOwnPropertyDescriptor(
