+++ tsconfig.json
@@ -10,7 +10,10 @@
     "outDir": "dist",
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
-    "noEmit": true
+    "noEmit": true,
+    "paths": {
+      "$sharedTypes/*": ["./types/*"]
+    }
   },
   "include": ["src", "forge.env.d.ts"],
   "exclude": ["gui"]
