+++ types/input-region.ts
@@ -1,0 +1,6 @@
+export interface InputRegion {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
