+++ types/update.ts
@@ -1,0 +1,6 @@
+export type UpdateInfo = {
+  version: string;
+  releaseNote: string;
+  url: string;
+  time: string;
+};
